requirements.txt
json-schema-matcher==0.1.7.1
six==1.16.0
bcrypt==4.0.1
pyjwt==2.8.0
flask~=3.0.2
websockets~=12.0
hypercorn==0.16.0